Implemented use of types other than int being supported by configs

+++ Agario 2/Source/Configs/PlayerConfigs.cs
@@ -5,6 +5,6 @@ public static class PlayerConfigs
     public static float StartingMaxSpeed = 500;
     public static float StartingRadius = 30;
     
-    public static int DashSpeedMultiplier = 4;
+    public static float DashSpeedMultiplier = 4;
     public static int DashSpanFrames = 20;
 }
